+++ CHANGELOG.md
@@ -2,7 +2,19 @@
 
 All notable changes to `alf-php-basic` will be documented in this file.
 
-## 1.0.0 - 2020-03-02
+## 1.0.1 - 2022-02-09
+
+- BugFix Changelog-Date for 1.0.0
+- ToDos for AlfBasicClass
+- Basic-Class: AlfBasicSingleton
+- Service: AlfProgramming
+- Basic-Class: AlfAttribute
+- Attribute: AlfAttrAutoComplete
+- Basic-Class: AlfType
+- Basic-Class: AlfTypeScalar
+- Remove test: AlfBasicClassTest
+
+## 1.0.0 - 2022-02-03
 
 - initial release
 - Alf\AlfPhp\AlfBasicClass